# backend/docs/swagger.yaml
definitions:
...
  domain.Prediction:
    properties:
      created_at:
        type: string
      explanation:
        type: string
      id:
        type: integer
      model_version:
        type: string
      part_category:
        type: string
      part_code:
        type: string
      part_name:
        type: string
      predicted_next_date:
        type: string
      predicted_next_mileage:
        type: integer
      probability:
        type: number
      recommendation:
        type: string
      remaining_days:
        type: integer
      remaining_km:
        type: integer
      risk_level:
        $ref: '#/definitions/domain.RiskLevel'
      risk_score:
        type: integer
      source:
        $ref: '#/definitions/domain.PredictionSource'
      vehicle_id:
        type: integer
    type: object
  domain.PredictionSource:
    enum:
    - rule_based
    - mock
    - ml_service
    type: string
    x-enum-varnames:
    - PredictionSourceRuleBased
    - PredictionSourceMock
    - PredictionSourceMLService
  domain.RiskLevel:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - RiskLevelLow
    - RiskLevelMedium
    - RiskLevelHigh
...
  handler.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  handler.predictionsResponse:
    properties:
      predictions:
        items:
          $ref: '#/definitions/domain.Prediction'
        type: array
      vehicle_id:
        type: integer
    type: object
  handler.pushPredictionsRequest:
    properties:
      predictions:
        items:
          $ref: '#/definitions/domain.Prediction'
        type: array
    required:
    - predictions
    type: object
...
info:
  contact: {}
paths:
...
  /api/internal/vehicles/{id}/predictions:
    post:
      consumes:
      - application/json
      description: Internal endpoint (no auth) for ML service to push predictions
        directly
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Predictions to save
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.pushPredictionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.predictionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Push predictions from ML service
      tags:
      - predictions-internal
...
  /api/vehicles/{id}/predictions:
    get:
      description: Returns latest predictions for vehicle parts. If predictions do
        not exist, backend generates and stores them.
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.predictionsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth: []
      summary: Get vehicle predictions
      tags:
      - predictions
  /api/vehicles/{id}/predictions/refresh:
    post:
      description: Calls ML service to regenerate predictions for the vehicle, saves
        results, and returns updated predictions
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.predictionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Refresh vehicle predictions
      tags:
      - predictions
...
swagger: "2.0"